Rename projector to EventPlaceHistoryProjector
